refactor(api): move attach-organization controller in bounded context
